media-source/URL-createObjectURL-step2.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>insertAdjacentHTML</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
test(function() {
    var mediaSource = new MediaSource();
    var url = window.URL.createObjectURL(mediaSource);
    assert_not_equals(url, null);
}, "URL.createObjectURL : 2. Return a unique MediaSource object URL that can be used to dereference the mediaSource argument, and run the rest of the algorithm asynchronously.");
</script>
</body>
</html>
